msm: Use 64 bit type for sclk
'us' can overflow and can potentially cause unexpected wakeup.

Change it to uint64_t.

+++ arch/arm/mach-msm/lpm_levels.c
@@ -354,7 +354,7 @@ static void lpm_system_prepare(struct lpm_system_state *system_state,
 	struct lpm_system_level *lvl;
 	uint32_t sclk;
 	int ret;
-	int64_t us = (~0ULL);
+	uint64_t us = (~0ULL);
 	int dbg_mask;
 	struct cpumask nextcpu;
 
@@ -392,7 +392,7 @@ static void lpm_system_prepare(struct lpm_system_state *system_state,
 	}
 
 	do_div(us, USEC_PER_SEC/SCLK_HZ);
-	sclk = (uint32_t)us;
+	sclk = us;
 	msm_mpm_enter_sleep(sclk, from_idle, &nextcpu);
 	if (from_idle)
 		time = ktime_to_ns(ktime_get());
